+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0]() - 2025-06-22
 ### Added 
-- OpenAI integration
+- OpenAI integration (Supports all services that comply with the OpenAI API spec. Requires the Responses API for structured output. Continue to use the separate Ollama provider until they add the Responses API. See https://github.com/ollama/ollama/issues/9659)
 - Jellyseer integration
+- Overseerr integration
 
 ### Changed
-- OpenAI provider now supports custom base URLs for API proxying.
+- OpenAI provider now supports custom base URLs for API proxying. 
 - Refactored Radarr and Sonarr into Providers of type "request"
 
 ## [1.1.2]() - 2025-06-21